Move classes to appropriate modules
